使用 __init_subclass__ 自动记录模型和框架

+++ build/build_delcache.py
@@ -1,5 +1,14 @@
 from pathlib import Path
-for cache in Path("src").glob("**/__pycache__"):
-    for file in cache.iterdir():
+import sys
+version = "cpython-" + "".join(map(str, sys.version_info[0:2]))
+for file in Path("src").glob("**/__pycache__/*.pyc"):
+    stem = file.stem
+    dot = stem.find(".")
+    file_name = stem[:dot]
+    file_info = stem[dot + 1:]
+    if file_info != version:
         file.unlink()
-    cache.rmdir()
+        continue
+    if not (file.parent.parent / f"{file_name}.py").exists():
+        file.unlink()
+        continue
